Remove CMake module usage
Since there will only ever be one "module", i.e. libssh2, for this super build,
it seems excessive to have the 10-ish lines of CMake in a separate file. The
project structure becomes essentially a single CMakeLists.txt file and an
archive of the libssh2 source code. It does not get much simplier than that.